rename: shoppingCart로 네이밍 변경
